requirements.txt
einops==0.6.0
omegaconf==2.2.3
opencv_python==4.6.0.66
pandas==1.5.2
Pillow==10.0.1
pytorch_lightning==1.8.3
scikit-learn==1.2.2
scipy==1.10.1
scikit-image==0.19.3
torch==1.13.1
torchmetrics==0.10.3
torchvision==0.14.1
tqdm==4.64.1
wandb==0.13.5
